Add high resolution timers for queue depth
This adds a Timer::HiRes call for every request we get, and every invocation of the 'queues' management command.  This should be negligible overhead for us to have millisecond quality data on how long requests have been enqueued for.

Note that I did not convert functionality such as the queue_relief function, or the Queues plugin.  StumbleUpon is not using those right now.